Started implementing the rules of the game. See description for more details
I made a class for rules. I have two rules so far. You have to play a
different type of card in order for it to be a valid play.  There is
also a method that checks to see if you've scored or not.
I also made a fragment that asks if you want to discard the card you're
about to play or if you want to play it.
If it's not a valid play, this dialog does not come up